feat: Add OZGCloudDataFormatService for form data processing and enhance display value methods in input fields
